refactor: remove RepayModal component and integrate repay functionality directly into Borrow page
style: adjust layout of Assets page for better responsiveness

feat: enhance Borrow page with embedded borrow and repay forms, including risk assessment and transaction fee validation

fix: swap BorrowingSection and AssetsList in Dashboard for improved user experience
