+++ mercata/ui/src/components/Positions.tsx
@@ -9,8 +9,6 @@ import { formatBalance } from "@/utils/numberUtils";
 interface BorrowingSectionProps {
   userCollaterals: CollateralData[];
   loanData: NewLoanData;
-  handleBorrow: () => void;
-  handleRepay: () => void;
 }
 
 // Reusable InfoTooltip component
@@ -28,7 +26,7 @@ const InfoTooltip = ({ children, content }: { children: React.ReactNode; content
   </Tooltip>
 );
 
-const PositionSection = ({ userCollaterals, loanData, handleBorrow, handleRepay }: BorrowingSectionProps) => {
+const PositionSection = ({ userCollaterals, loanData }: BorrowingSectionProps) => {
 
   function getTextColor(value: number, maxValue = 10) {
   const clamped = Math.min(Math.max(value, 1), maxValue);
@@ -42,50 +40,26 @@ const PositionSection = ({ userCollaterals, loanData, handleBorrow, handleRepay 
   
   return (
     <Card className="border border-gray-100 shadow-sm">
-      <CardHeader className="flex flex-row items-center justify-between pb-2">
+      <CardHeader className="pb-4">
         <div>
           <CardTitle className="text-2xl font-bold">Your Position</CardTitle>
         </div>
-        <div className="flex gap-2">
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <Button onClick={handleBorrow} className="flex items-center gap-2">
-                Borrow
-              </Button>
-            </TooltipTrigger>
-            <TooltipContent>
-              <p>Borrow USDST against your supplied collateral. You'll need to supply tokens first to enable borrowing.</p>
-            </TooltipContent>
-          </Tooltip>
-          <Tooltip>
-            <TooltipTrigger asChild>
-              <Button onClick={handleRepay} className="flex items-center gap-2">
-                Repay
-              </Button>
-            </TooltipTrigger>
-            <TooltipContent>
-              <p>Repay your borrowed USDST to reduce debt and improve your position's health factor.</p>
-            </TooltipContent>
-          </Tooltip>
-        </div>
       </CardHeader>
-      <CardContent>
-        <div className="py-4">
-          <div className="space-y-6">
-
-            {/* Added extra spacing with mt-8 to separate indicators from data */}
-            <div className="grid grid-cols-2 md:grid-cols-5 gap-x-8 gap-y-6 mt-8">
-              <div className="flex flex-col">
-                <span className="text-gray-600">Total Borrowed</span>
-                <span className="font-semibold">
+      <CardContent className="px-6 pb-8">
+        <div className="py-6">
+          <div className="space-y-8">
+            <div className="flex flex-col gap-4">
+              <div className="flex flex-col space-y-3 p-4 bg-gray-50 rounded-lg">
+                <span className="text-gray-600 text-sm font-medium">Total Borrowed</span>
+                <span className="font-semibold text-lg">
                   {formatBalance(loanData?.totalAmountOwed || 0n, "USDST", 18, 2, 2)}
                 </span>
               </div>
-              <div className="flex flex-col">
+              <div className="flex flex-col space-y-3 p-4 bg-gray-50 rounded-lg">
                 <InfoTooltip content="Measures your position's safety. Higher is better. Below 1.0 means risk of liquidation. No loan means you have no outstanding debt.">
-                  <span className="text-gray-600">Health Factor</span>
+                  <span className="text-gray-600 text-sm font-medium">Health Factor</span>
                 </InfoTooltip>
-                <span className="font-semibold text-green-500" style={{ color: getTextColor((loanData?.healthFactor)) }}>
+                <span className="font-semibold text-lg" style={{ color: getTextColor((loanData?.healthFactor)) }}>
                   {(() => {
                     // Check if there's no outstanding debt
                     const totalAmountOwed = loanData?.totalAmountOwed ? parseFloat(formatUnits(loanData.totalAmountOwed.toString(), 18)) : 0;
@@ -100,27 +74,27 @@ const PositionSection = ({ userCollaterals, loanData, handleBorrow, handleRepay 
                   })()}
                 </span>
               </div>
-              <div className="flex flex-col">
+              <div className="flex flex-col space-y-3 p-4 bg-gray-50 rounded-lg">
                 <InfoTooltip content="You need to supply tokens as collateral before you can borrow. Click 'Supply' in the Eligible Collateral table below to get started.">
-                  <span className="text-gray-600">Available Borrowing Power</span>
+                  <span className="text-gray-600 text-sm font-medium">Available Borrowing Power</span>
                 </InfoTooltip>
-                <span className="font-semibold">
+                <span className="font-semibold text-lg">
                   {formatBalance(loanData?.maxAvailableToBorrowUSD || 0n, "USDST", 18, 2, 2)}
                 </span>
               </div>
-              <div className="flex flex-col">
+              <div className="flex flex-col space-y-3 p-4 bg-gray-50 rounded-lg">
                 <InfoTooltip content="Accumulated interest on your borrowed amount. This increases over time and should be repaid along with your principal.">
-                  <span className="text-gray-600">Interest Owed</span>
+                  <span className="text-gray-600 text-sm font-medium">Interest Owed</span>
                 </InfoTooltip>
-                <span className="font-semibold">
+                <span className="font-semibold text-lg">
                   {formatBalance(loanData?.accruedInterest || 0n, "USDST", 18, 2, 2)}
                 </span>
               </div>
-              <div className="flex flex-col">
+              <div className="flex flex-col space-y-3 p-4 bg-gray-50 rounded-lg">
                 <InfoTooltip content="Annual percentage rate you pay on borrowed amounts. This rate applies to your total borrowed amount.">
-                  <span className="text-gray-600">Interest Rate</span>
+                  <span className="text-gray-600 text-sm font-medium">Interest Rate</span>
                 </InfoTooltip>
-                <span className="font-semibold">{loanData?.interestRate || 0}%</span>
+                <span className="font-semibold text-lg">{loanData?.interestRate || 0}%</span>
               </div>
             </div>
           </div>
